Moved method post_user_data from test_auth_endpoints to base_case
